Display warning when a suspected reference is loaded
To properly load references a custom deserializer is needed. Otherwise we do not know what object it is refering to, just that it is refering to _something_.

I guess we do kinda know: We know the object id used and we know the expected type. So if we somehow manage to connect id with objects we are actually good!